// src/models/ResponseValue.ts
export interface ResponseValue {
  "@odata.type": string;
  "@odata.id": string;
  "@odata.etag": string;
  "@odata.editLink": string;
  Id: number;
  Title: string;
  IconName: string;
  QuickLinkUrl: string;
  Display: boolean;
  ForeGroundColor: string;
  BackgroundColor: string;
  ID: number;
  ItemOrder: number;
  BackgroundImageURL?:string;
}